+++ src/main/kotlin/org/pcsoft/framework/kunit/distance/KDistanceAreaUnit.kt
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) KleinerHacker alias Pfeiffer C Soft 2026.
+ * This work is licensed under the Apache License, Version 2.0.
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at:
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, this software is distributed on an “AS IS” BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and limitations.
+ */
+
+package org.pcsoft.framework.kunit.distance
+
+/** Bare area-unit token for the area prefix `infix` DSL (e.g. `5 kilo squareMeters`). */
+@JvmInline
+value class KDistanceAreaUnit internal constructor(internal val unit: KDistanceUnit)